m3-9 fixup: move authResult to auth_helpers.go (CODE-7 boundary inversion)
